Fixed msvc warnings

+++ include/jcdp/jacobian.inl
@@ -36,13 +36,13 @@ inline auto Jacobian::generate_random(
             }).value();
    jac.m = size_distribution(gen);
 
-   jac.kl = std::round((jac.m - 1) * density_distribution(gen));
-   jac.ku = std::round((jac.n - 1) * density_distribution(gen));
+   jac.kl = static_cast<std::size_t>(std::round((jac.m - 1) * density_distribution(gen)));
+   jac.ku = static_cast<std::size_t>(std::round((jac.n - 1) * density_distribution(gen)));
 
    const std::size_t max_mn = std::max(jac.m, jac.n);
    jac.non_zero_elements = max_mn;
-   jac.non_zero_elements += std::round(
-        (max_mn - jac.m * jac.n) * density_distribution(gen));
+   jac.non_zero_elements += static_cast<std::size_t>(std::round(
+        (max_mn - jac.m * jac.n) * density_distribution(gen)));
 
    jac.edges_in_dag = dag_size_distribution(gen);
    jac.tangent_factor = tangent_factor_distribution(gen);
